requirements.txt
pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.3.0
streamlit>=1.28.0
plotly>=5.17.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.25.0
anthropic>=0.39.0
python-dotenv>=1.0.0
